Moves types into its own package for goMaker
