+++ src/dev_tools/goMaker/types/codebase.go
@@ -1,4 +1,4 @@
-package main
+package types
 
 import (
 	"encoding/json"
