+++ src/dev_tools/goMaker/types/codebase_load.go
@@ -1,4 +1,4 @@
-package main
+package types
 
 import (
 	"sort"
