+++ src/dev_tools/goMaker/types/command.go
@@ -1,4 +1,4 @@
-package main
+package types
 
 import (
 	"strings"
