+++ src/dev_tools/goMaker/types/csv.go
@@ -1,4 +1,4 @@
-package main
+package types
 
 import (
 	"os"
