+++ src/dev_tools/goMaker/types/endpoint.go
@@ -1,4 +1,4 @@
-package main
+package types
 
 import (
 	"strings"
