+++ src/dev_tools/goMaker/types/option_template.go
@@ -1,4 +1,4 @@
-package main
+package types
 
 import (
 	"bytes"
